test(select): add test cases for selects with blank disabled options
An earlier commit 1274c4000323a5b6c72f1168afd1a19c0233f73d caused an error where the first option of
a select would be skipped over if it had a blank disabled value. These tests demonstrate that with
that commit in place, blank disabled options are skipped in a select. When the commit is reverted,
the correct behavior is seen that the blank disabled option is still selected in both selects
marked with required and those that have optional choices.

Relates to #7715